django/requirements.txt
Django
psycopg2-binary
djangorestframework
markdown
django-filter
django-cors-headers
python-dateutil
coverage
django-nose
nose
